refactor(test): group remaining inference and agent coverage
