+++ test/agents/hermes/reviewed-hermes-platform-action.test.ts
@@ -9,7 +9,7 @@ import path from "node:path";
 import { describe, expect, it } from "vitest";
 import YAML from "yaml";
 
-const ROOT = path.resolve(import.meta.dirname, "..");
+const ROOT = path.resolve(import.meta.dirname, "../../..");
 const ACTION_PATH = path.join(
   ROOT,
   ".github",
