+++ test/agents/hermes/seed-hermes-dashboard-config.test.ts
