+++ test/agents/openclaw/kimi-inference-compat-plugin.test.ts
