+++ test/inference/managed/nemotron-inference-fix.test.ts
@@ -8,7 +8,7 @@ import os from "node:os";
 import path from "node:path";
 import { describe, expect, it } from "vitest";
 
-const START_SCRIPT = path.join(import.meta.dirname, "..", "scripts", "nemoclaw-start.sh");
+const START_SCRIPT = path.join(import.meta.dirname, "../../..", "scripts", "nemoclaw-start.sh");
 const NEMOTRON_FIX_SOURCE = path.join(
   import.meta.dirname,
   "..",
